Fix the 32-bit build of the memory limit error
math.MaxInt64 is an untyped constant, so passing it straight to
fmt.Errorf gave it its default type of int. On a 32-bit target such as
armhf that overflows and the package no longer compiles:

  cannot use math.MaxInt64 (untyped int constant 9223372036854775807)
  as int value in argument to fmt.Errorf (overflows)

amd64 and arm64 were unaffected, so go build, go test and the build
workflow all passed, and only the cross-compile in the publish workflow
caught it.

Convert to int64 explicitly. Verified by building every target the
publish workflow produces, and by type-checking the tests for armhf.

+++ commands/local_run_resources.go
@@ -135,7 +135,9 @@ func negativeMemoryError(value string) error {
 }
 
 func oversizedMemoryError(value string) error {
-	return fmt.Errorf("invalid memory value %q: must be under %d bytes", value, math.MaxInt64)
+	// Typed as int64 so the constant does not take its default type of int,
+	// which overflows when cross-compiling for a 32-bit target such as armhf.
+	return fmt.Errorf("invalid memory value %q: must be under %d bytes", value, int64(math.MaxInt64))
 }
 
 // toDockerCPU converts a Kubernetes CPU quantity such as "100m" (millicores)
